requirements.txt
openai>=1.3.0
PySimpleGUI>=4.60.0
python-dateutil>=2.8.0